tp6

+++ tp6/tp6.h
@@ -37,7 +37,7 @@ typedef struct
 
 
 int lire_deux_octets(int fd, uint16 *val);
-int lire_quatre_octets(int fd, uint16 *val);
+int lire_quatre_octets(int fd, uint32 *val);
 int lire_entete(int de, entete_bmp *entete);
 
 int ecrire_deux_octets(int fd, uint16 *val);
